Add cache_dir config, default to /tmp/VSB
Currently downloaded datasets are stored in the current working
directory.

This can result in multiple copies of the same dataset being
downloaded; which also slows down startup (as we end up re-downloading
even if the files already exist somewhere locally.

Change so we use a fixed location, defaulting to /tmp/VSB/cache.