Perf: avoid loading workload for MasterRunners
locust has three concrete Runner types: LocalRunner, WorkerRunner and
MasterRunner. The first two are actual runners where tasks are run and
work performed, the latter is used to coordinate a distributed run
(when --processes is used).

We don't actually need to load the workload on the MasterRunner - it
never runs any tasks itself. As such for large datasets we can save
the potentially non-trivial amount of memory needed to load the
dataset by not loading it on the MasterRunner.