Fix multiprocess worker heartbeat issues with large dataset loads
When running in multiprocess mode if the dataset is large (takes >1s
to load from disk) then worker heartbeats can be missed and the master
locust instance ends up terminating the workers:

    [2024-05-08 09:25:14,348] Delirium.local/INFO/locust.runners: Worker Delirium.local_c7ab179ec332423da67429fb8f0a8272 failed to send heartbeat, setting state to missing.
    [2024-05-08 09:25:14,380] Delirium.local/INFO/locust.runners: Worker Delirium.local_18d8b48402c94f91bdc2d1a878327ac8 failed to send heartbeat, setting state to missing.
    [2024-05-08 09:25:14,382] Delirium.local/INFO/locust.runners: The last worker went missing, stopping test.

This is due to dataset loading running in a gevent greenlet, and the
loading into a Pandas DataFrame can block the event loop for longer
than the heartbeat interval (1s). As such, the heartbeat request
message from the master cannot be responded to and the master has it's
timeout expire.

Fix this by running the worker dataset loading in a separate (real)
thread, which does not interfere with the greenlets.