Move grpc_event monkey-patching to pinecone.py
This was originally done in locustfile, however it appears that some
changes to import order resolution mean that grpc is not patched
sufficiently early. This manifests as a hang / deadlock seen when
loading large workloads (e.g. nq768) with multiple processes.

Move to pinecone.py which is where we use gRPC.