Use pyarrow.dataset streaming API to reduce memory usage
For parquet file based datasets, the records to be populated into the
database are all loaded into memory at once (into a pandas DataFrame),
then chunked into num_users chunks (so each user could independently
upsert their own chunk)

This required each locust process to (temporarily) have the entire
dataset in memory (when converted into a pandas DataFrame) - which can
be a significant amount of RAM for larger datasets. For example, the
nq-768-tasb dataset is around 8GB in size, so each process consumed in
excess of 8GB when loading the dataset - around 10GB on macOS.

This significantly limits the number of locust processes which are
usable on a typical h/w setup.

To reduce the memory footprint we make use of pyarrow's Dataset
streaming API, which allows a dataset of multiple Parquet files to be
read in chunks, without loading the entire set into RAM. This does
however rely on a dataset consisting of multiple files (we can only
chunk at the whole file level with this API), however _most_ large
datasets are spread over multiple files...

Note there is one situation where we cannot make this optimisatio -
when the dataset to be loaded has been limited in record count
(e.g. for the -test workloads) - we cannot use the streaming API any
only read some of the file - at least not in a straightforward
manner. As such, continue to use the old approach of reading the
(limited) set of records into memory, then splitting that into chunks.