Replace tqdm with rich.progress
tqdm's progress interacts poorly with gevent - when it is used to
display the download progress, VSB hangs after the run is completed;
sending a SIGINT shows that tqdm has a thread stuck waiting to acquire
a lock:

    ^CKeyboardInterrupt
    2024-05-09T13:31:22Z
    Exception ignored in atexit callback: <bound method TMonitor.exit of <TMonitor(Thread-4, stopped daemon 6050743936)>>
    Traceback (most recent call last):
      File /Users/dave/repos/pinecone-io/VSB/.venv/lib/python3.11/site-packages/tqdm/_monitor.py, line 44, in exit
	self.join()
      File /opt/homebrew/Cellar/python@3.11/3.11.9/Frameworks/Python.framework/Versions/3.11/lib/python3.11/threading.py, line 1119, in join
	self._wait_for_tstate_lock()

Another progress bar (rich) doesn't have this problem, and for bonus
points is more colorful. Replace tqdm with rich.progress.