Add YFCC Parquet workload (filtered)
Add workloads 'yfcc-10M' and yfcc-test (100k documents) from
https://big-ann-benchmarks.com/neurips23.html 'Filtered Search'
workload.

This dataset includes metadata (bag of integer tags) on each record,
and metadata filtering as part of the query set, so we need to expand
the Parquet file loader to read metadata column and include in upsert;
then read the filter and include in the search requests.

Additionally YFCC has uint8_t embeddings, so we need to support
integer or float types of vector.