Testing: Move Pinecone index creation into pytest fixture
With the expansion of the number of workloads tested, we need to make
use of multiple Pinecone indexes to test against (different workloads
have different dimensions, and the dimension of a Pinecone index is
fixed at index creation time).

For integration tests, move the creation of indexes from a separate
GitHub actions script into a pytest fixture; which allows us to create
whatever indexes are needed for the tests being executed via pytest
Fixtures.