Users: Avoid multiple PopulateUsers from clearing index
Currently any existing index is cleared at the start of the Populate
phase (when the DB subclass' __init__ method is called). This is fine
if there is just a single user specified (--users=1), however if there
are multiple users (particulary running on multiple processes) then we
will clear the index multiple times, which may result in losing
records if the one PopulateUser starts upserting records before other
PopulateUser instance(s) have cleared the index.

Address this by introducing a new Setup phase and corresponding
SetupUser. This new phase will only call the index setup code on a
single task instance, waiting for it to complete before advancing to
the PopulatePhase (with multiple users).