Validate conditionally-required arguments at startup
Different databases require different arguments - for example:

    pinecone requires that --pinecone_api_key and --pinecone_index_name are specified.

Currently this is not validated as part of parsing the command-line
options, and will only be detected later on when we attempt to
initialise the given database, which isn't necessarily clear to the
user.

Improve this by adding explicit checking that such
conditionally-required arguments are indeed specified, printing out a
relevent error message if any are omitted.

Also improve the general format and layout of options help; listing
--workload and --database (the two most important arguments) in their
own group.