Add initial Metrics reporting infrastructure; calc Recall
Add support for gathering additional custom metrics on search results,
such as Recall, MRR, etc. This includes support for multi-process
runs, where each worker calculates metrics for the requests it runs,
and then they are sent back to the master runner for aggregation and
reporting at the end of the run.

In single-process mode, we simply call a new method
metrics.record_metrics() after performing a search request, passing in
the result of the request. record_metrics() will calculate metrics of
interest (currently just recall), and store the results in a global
dict (metric name -> values).

At the end of the run the metric values are printed to stdout.

In multi-process mode we don't want to be sending every result from
the workers to the master as the comms overhead would be significant
with many processes / users; instead we make use of locust's stats
reporting hooks, of which there are two principal ones:

- report_to_master: This event is triggered periodically (~1 Hz) on
  each worker instance when stats should be sent back to the master.

- worker_report: This event is triggered on the master instance when a
  new stats report arrives on the master.

This allows us to accumulate ~1s worth of metrics on each worker
runner, then send the entire batch to the master runner.

Initially this patch just adds the basic infrastructure to calculate
and record one metric (recall), simply recording every metric value
and printing them on exit. Subsequent patches will build on this to
calculate summaries (percentiles etc), add additional metrics (mean
reciprocal rank, mean average precision, ...),