Add vsb-specific logging, move locust messages to seperate log file
Locust's INFO level log messages contain a lot of information not
directly relevent to VSB, and it would be helpful to users if they had
some high-level information about the status of the run (what phase,
how much work done, results).

This patch splits the outputted log messages in two:

* Only VSB specific messages (via a new `vsb.logger`) are emitted to
  stderr, using the rich library to improve formatting, and are also
  sent to a new vsb.log file which is created per invocation.

* All other log messages are only sent to the vsb.log file.

This reserves the screen for just high-level information messages to
the user, with detailed log information saved to a file.

It also writes the JSON statistics information to log_dir/stats.json.