Add progress bars for main phases
Add progress bars for each of the main phases of an experiment -
Setup, Populate and Run. For Populate and Run we require the total
number of records / queries to show a useful end progress so far.

For the Populate phase we include the current rate of upsert. For the
Run phase we include the current latency and recall(*) values, these
require additional metrics of how many records have been upserted so
far (note that we generally upsert in batches, so the existing number
of Population requests is not sufficient.

(*) For recall we cannot calculate the current metric (last 10s), as
    we only have a single histogram accumulating the results - instead
    this shows the overall recall so far. There's an improvement
    raised to fix this (#108).