dataset: Fix bug where newly-downloaded files were not used
In Dataset.get_batch_iterator() with limit=0, the list of files in the
dataset was not updated after downloading them from the remote
bucket. The effect of this was if the dataset was not cached locally
prior to running vsb (i.e. first run), then no dataset files would be
found and the iterator would return nothing.

This was exposed when modifying test_dataset.py to remove a test which
essentially 'pre-fetched' the dataset files for a test using
get_batch_iterator().

Fix by moving the download step earlier, before we set the list of
parquet files.