Use multipart version of YFCC-10M
The version of the YFCC-10M dataset currently used is a single parquet
file; which means the populate phase cannot be parallelised.

Split the single file into multiple (100,000 records per file),
allowing concurrent upload. File split with:

    ./rewrite_parquet_file_split.py \
        /tmp/VSB/cache/yfcc-10M-filter-euclidean-formatted/passages/part-0.parquet \
	yfcc-10M-filter-euclidean-formatted-multipart 10000 100000

Also tweak how download progress bars are displayed - with a large
number of files they fill up the screen, so remove a task when the
files is downloaded.