Output metrics only on successful run
Previously the experiment metrics and stats file was written via the
'quitting' event, so would occur even on unclean exit (Ctrl-C, fatal
exception thrown).

This is both misleading (the stats are only partial for the run so
far), and verbose / confusing - we print a bunch of text when the user
Ctrl-C's the process.

Only emit metrics and perform stats file creation if the LoadShape
reached the Done phase of the experiment - after all executed phases
are correctly executed.