Add --requests_per_sec cmdline option to limit tput
Currently the Run phase just issues requests as quickly as
possible. To be able to run experiments for a wider range of workloads
we need to be able to support running at a specified limit (1 QPS, 10
QPS, ...).

This patch adds a new command line argument: --requests_per_sec=N,
which allows the oner to specify the _target_ requests per second
which VSB will issue in the Run phase.

* If there are multiple users configured (--users), then the target is
  split between them - e.g. for --users=4 and --request_per_sec=100,
  each user will attempt to issue 25 requests per second.

* Fractional values can be specified, allowing <1 request per second.

* If a value of 0 is specified (the default) then no limit is imposed
  - requests will be sent as fast as possible (sequentially for each
  user).

Fixes #125.