Execute each query only once when multi-process
Each process reads a complete copy of the queries dataset, which is
then distributed across all clients running in that process.

This means that with N processes, we will issue the same query N
times. This is undesirable as we essentially end up prefetching query
results; the second and subsequent execution of the same query will be
unrealistically faster as the result was just calculated.

Fix by modifying query assignment to be similar to Populate phase -
each RunUser is assigned a user_id, and uses that to request a slice
of queries for that given user.

Fixes #61.