pgvector: Set Docker RAM to 8GB and maintenance_work_mem to 4GB
During testing it has been observed that pgvector runs out of
maintenance_work_mem during index build. This is due to
maintenance_work_mem - the setting which controls how much RAM can be
used for index build - being still at the default value of 64MB. For
hnsw indxes this needs to be at least as large as the index - for
example even mnist's HNSW index is ~240MB.

As such, add a new argument for pgevector to allow
maintenance_work_mem to be set:

* --pgvector_maintenance_work_mem:

Set this to a consistent, reasonable default value of 4GB. Also limit
the pgvector Docker container to 8GB (and 2 cpu cores) so we have a
fixed environment to allow more reproducable performance numbers.