Remove Pydantic validation overhead when populating from Parquet files
A significant amount of load time is spent by pydantic validating each
record's values vector, element by element. This is unnecessary because
the Parquet files we ingest are homogeneous, but pydantic doesn't have
this context when we pass a regular list into the Record constructor.

Added a custom field validator that runs instead of Pydantic's default
for Record's values field. Instead of checking each element in values
(slow), we ensure the dtype of the NDArray passed in is integral or
floating (fast). If the Record constructor is called with a non-ndarray,
we default to Pydantic's original validator.

We see a 3.3x reduction in load time from 27s to 8s on local pgvector +
mnist, and 54s to 27s (2x faster) on pinecone + mnist. When compared to
YCSB (13s load on local pgvector + mnist), we are as good or faster :)