pgvector: Fix automatic ivfflat lists + skip_populate
As part of #150, `--pgvector_ivfflat_lists` was changed to default to
automatically selecting list count based on the number of
records. However, the value was calcuated in finalize_population(),
which is not called if `--skip_populate` is passed to VSB.

As such, the combination of using an pgvector with an ivfflat index
with automatic #lists resulted in failures at query time, as the
number of search candidates in turn is derived from ivfflat_lists:

    psycopg.errors.InvalidParameterValue: 0 is outside the valid range for parameter "ivfflat.probes" (1 .. 32768)

Fix by hoisting the calculation of the ivfflat_lists to the __init__
method, so it is always available even if population is skipped.