Show indeterminate tasks' progress as pulsed bar
For tasks which we cannot measure their incremental progress until
complete (e.g. we issue an SQL statement and wait until it returns),
we currently show a progress bar of '0/1'. This isn't doesn't really
show to the user that something _is_ happening.

To make it more explicit that a task is running, make use of the
rich.progress 'pulse' progress bar style, where an animation is shown
in the progress bar. This is done by setting the tasks' total to
None. When the task actually does complete, update the total (and
completed) to the value 1, so we continue to show a progress of '1/1'.