Add test workload filtering and full loading The iterator returned by get_record_batch_iter would still return multiple batches in high-count test workloads (e.g. nq_768), but we only took the first batch. Now all batches are loaded.
Workloads like yfcc incorporated metadata filtering on tags, which
also led to inaccurate simulated results with just top-k distance.
Metadata filtering will now occur if yfcc metadata fields are
detected on the test records.

There also exists edge cases in test workloads that involve a subset
of records. Some queries will filter on tags that are completely not
present in the test dataset. This leads to both an actual and expected
response of an empty list. We consider this as 0 recall, but should
consider this as 100% recall. (The answer was [], and we returned []).