WIP: Change to scikit-lean's NearestNeighbors model
This avoids a brute-force search in favour of scikit-learn
NearestNeighbors model.

This gives significant improvements again for cosine and euclidean
distance (as we can use it's native distance function):

The improvement is less significant (TODO: measure) for dot-product as
that's not supported natively and hence we have to pass our own
function, which appears to be more expensive than the built-ins.

Note this is currently ~the same speed as the previous version (bulk
recalc) for mnist:

    --------------------------------------------------------------------------------------------- benchmark: 2 tests --------------------------------------------------------------------------------------------
    Name (time in ms)                                   Min               Max              Mean            StdDev            Median               IQR            Outliers       OPS            Rounds  Iterations
    -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
    test_topk_recalc_mnist_euclidean (0017_bulk)     2.3764 (1.0)      3.3675 (1.0)      2.6242 (1.0)      0.1281 (1.32)     2.6046 (1.0)      0.1778 (1.56)         69;3  381.0616 (1.0)         272           1
    test_topk_recalc_mnist_euclidean (NOW)           2.6085 (1.10)     3.4282 (1.02)     2.7778 (1.06)     0.0971 (1.0)      2.7648 (1.06)     0.1137 (1.0)          52;5  359.9985 (0.94)        245           1
    -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

We *might* see more significant speedup for larger datasets and/or
different metrics, but that requires additional benchmarks (perhaps
using randomised data so we have a the same dataset for each metric?)