pgvector: Add support for running without an index
If an pgvector index is not created, then Postgres will perform an
exact, exhaustive search. This is typically (much) slower than using
an index, but it does guarantee perfect recall.

As such, this can be a useful mode for development / testing datasets,
or even establishing a baseline of how Postgres behaves when no index
is used.

Add support for not running with an index by adding the 'none' option
to --pgvector_index_type.