Add WorkloadSequence functionality (multiple workload support)
This commit introduces a new class VectorWorkloadSequence, that
wraps a sequence of VectorWorkloads. This change allows us to offer
workloads that use multiple datasets and/or more complex sequences
of operations.

Large API changes were made to LoadShape, Locust Users, and dataset
loading. In summary, LoadShape and Users were refactored to iterate
over a WorkloadSequence with the help of a new class IterationHelper
to manage distributed iteration. Old singular workloads retain
functionality with the help of a SingleVectorWorkloadSequence class
that automatically wraps single VectorWorkloads; the process of
adding a new single workload remains the same.