Rework to pull-based iteration
It was a little too invasive for the MasterRunner to push an
iteration update and wait for a response from each worker. We
simplify the iteration tracker to remove IterationHelper, and
instead introduce a Subscriber class that's updated from the
master and can be pulled by workers at any time. This stores
an integer "iteration" index, which workers use to pick
workloads from the sequence.