Fix race conditions, tests
I might need a few days in the psych ward after this one.

1. Single-process runs with LocalRunner were hanging sometimes, the
LoadShape's tick() wouldn't be called if it wasn't called before
environment setup happened. Switch environment setup to be event-
based, triggered by the LoadShape reaching the WaitingForWorkers
state.

2. ^ That change broke multi-process runs, and I spent a bunch of
time realizing that for some reason, the logger.debug statements
I put in tick() to find the deadlock actually were the deadlock.
logger.debug() uses a semaphore, and in forked processes, that
lock might be forked as locked and never reopened. So the lesson
was to avoid using logging with multiprocessing and forking.