Add --overwrite to prevent accidental deletion of existing indexes (#262)
## Problem

VSB automatically deletes any existing data before populating (if
`--skip_populate` is not specified). This makes it simple to run a
complete benchmark irrespective of the current status of the index (we
start with a clean, consistent state), but this can be undesirable and
costly if the existing data is large and the user forgot to specify
`--skip-populate`).

## Solution

As such, require the user to specify the new flag `--overwrite` before
deleting any existing index. If the user doesn't specify --overwrite
then the run is aborted:

2024-11-01T09:28:08 CRITICAL PineconeDB: Index 'vsb-mnist-test' already
exists - cowardly refusing to overwrite existing data.Specify
--overwrite to delete it, or specify --skip-populate to skip population
phase.

## Type of Change

- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Test Plan

Describe specific steps for validating this change.