Avoid logging 2 copies of error+ log messages (#264)
## Problem

Currently log messages logged to vsb.logger at ERROR or greater are
printed twice to stdout.

This is due to the current logger setup. Currently we:

1. We configure the root logger to log all enabled messages to a file
   (vsb.log), additionally logging error+ log messages to stdout.

2. We configure a named 'vsb' logger to also log all enabled messages
   the same file (vsb.log), additionally logging all enabled messages
   (default INFO+) to stdout.

(This is to avoid printing the somewhat noisy Locust INFO messages to
stdout, but still recording to a file in case they are needed for
debugging; and then just showing the VSB messages we added to stdout.

However, by default Python loggers propogate log messages to the
parent, so the VSB logger prints ERROR+ messages itself, then passes
to the parent (root) logger which logs a second time.

## Solution

Fix this by disabling propogation from vsb logger to root; then
explicilty adding the file logger handler to vsb logger so we still
see VSB's log messages in the file.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)

## Test Plan

Describe specific steps for validating this change.