Make TestSubsetWorkloadKNN use non-equal values
In TestSubsetWorkloadKNN, not all distances between each record and
given query vector are unique (for a given distance metric / query).

This meant the test assertion was brittle as it depended on the order
that two identical values were sorted.

Fix the test by giving all records so the expected top_k will always
be correct.