Fix issues with --synthetic_metadata parsing, expand examples (#289)
## Problem

`--synthetic_metadata` was not parsed correctly if multiple instances of
synthetic_metadata were passed, due to the use of lambdas in the
metadata generator parser - they captured the initial value of the
arguments passed, which had the effect of subsequent
--synthetic_metadata fields using the "wrong" values.

## Solution

Fix by replacing these with normal functions; then add tests to cover
each metadata field type.

Also expand the documentation for synthetic metadata, hopefully make it
easier for users to make use of.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)