CI: Use pyproject version of black (#309)
## Problem

Prior to this if black@stable version updates (e.g. to black v25) then
we end up using a different version of black for GitHub CI action vs
local development (which is still using black 24).

## Solution

Set `use_pyproject: true` so the versions are in sync.

## Type of Change

- [x] Infrastructure change (CI configs, etc)

## Test Plan

Describe specific steps for validating this change.